move client to separate dir
