+++ tesp_api/service/client/__init__.py
@@ -1,0 +1,5 @@
+from tesp_api.service.client.client_s3 import ClientS3
+from tesp_api.service.client.client_ftp import ClientFTP
+
+client_s3 = ClientS3()
+client_ftp = ClientFTP()
